+++ BLL/Services/StewardessService.cs
@@ -22,6 +22,11 @@ namespace BLL.Services
 
         public StewardessDTO GetById(int id)
         {
+            if (id < 0)
+            {
+                throw new ArgumentException();
+            }
+            
             var item = mapper.Map<Stewardess, StewardessDTO>(unitOfWork.StewardessRepository.Get(id));
 
             if (item == null)
@@ -59,6 +64,11 @@ namespace BLL.Services
 
         public void Remove(int id)
         {
+            if (id < 0)
+            {
+                throw new ArgumentException();
+            }
+            
             unitOfWork.StewardessRepository.Delete(id);
         }
